Move encode.go contents to oauth.go
